% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grazing_calc.R
\name{frgd_}
\alias{frgd_}
\title{Calculate forage demand}
\usage{
frgd_(
  head,
  wt,
  days,
  area,
  prct_int = 2,
  input_system = "imperial",
  output_unit = NULL
)
}
\arguments{
\item{head}{Number of head}

\item{wt}{Mean animal weight}

\item{days}{Number of days in the pasture}

\item{area}{Area of the pasture}

\item{prct_int}{Estimated daily dry matter intake expressed on a percent body weight basis. Common values range between 1.5 and 3 percent. The default is 2 percent.}

\item{input_system}{Unit system for \code{wt} and \code{area}; either \code{"imperial"} or \code{"metric"}. If \code{"imperial"}, assumes \code{"lb"} and \code{"ac"}. If \code{"metric"}, assumes \code{"kg"} and \code{"ha"}. Default is \code{"imperial"}. Respects inherited units.}

\item{output_unit}{Optional. Desired output units. Typically either \code{"lb/ac"} or \code{"kg/ha"}. Default matches the input unit system.}
}
\description{
Calculate pasture forage demand expressed in "lb/ac" or "kg/ha"
}
\examples{
frgd_(head = 136, wt = 1350, days = 45, area = 931, prct_int = 2.5, input_system = "imperial")

frgd_(55, 480, 30, 60, input_system = "metric")

}
